Fix bjork.

+++ src/diffusion_for_multi_scale_molecular_dynamics/sample_diffusion.py
@@ -124,7 +124,7 @@ def main(args: Optional[Any] = None):
     if 'constrained_sampling' in hyper_params:
         logger.info("Constrained Sampling is activated")
         constraint_dict = hyper_params['constrained_sampling']
-        constrained_atom_indices = torch.Tensor(constraint_dict['constrained_atom_indices'])
+        constrained_atom_indices = torch.Tensor(constraint_dict['constrained_atom_indices']).to(torch.int64)
         cif_file_path = Path(args.path_to_constraint_cif_file)
         assert cif_file_path.is_file(), "The constraint cif file does not exist."
         logger.info(f"Constrained cif file is {cif_file_path}")
